feat: add utils.cpp
